# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Hoteles",
        "description": "API para gestionar hoteles, habitaciones, huéspedes y servicios",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/room": {
            "get": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Obtener lista de habitaciones",
                "description": "Devuelve una lista de habitaciones con la posibilidad de filtrar por número, tipo, precio por noche e id del hotel, y paginar los resultados.",
                "operationId": "1758cfa0a4f31164f32a8c2192f97d86",
                "parameters": [
                    {
                        "name": "number",
                        "in": "query",
                        "description": "Filtrar por número de la habitación",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrar por el tipo de la habitación",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nightPrice",
                        "in": "query",
                        "description": "Filtrar por precio de la habitación",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hotelId",
                        "in": "query",
                        "description": "Filtrar por id del hotel",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de habitaciones obtenida"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            },
            "post": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Crear una nueva habitación",
                "description": "Crea una nueva habitación con los datos introducidos",
                "operationId": "9be38be1eed9b9d7e62fd45fca294b7c",
                "requestBody": {
                    "description": "Datos necesarios para crear una habitación",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "number": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "nightPrice": {
                                        "type": "integer"
                                    },
                                    "hotelId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Habitación creada"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/room/{id}": {
            "get": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Obtener una habitación",
                "operationId": "c45e99f9ccf62d45114462c46ea5036a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Habitación encontrada"
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Actualizar una habitación",
                "description": "Actualiza una habitación con los datos introducidos",
                "operationId": "096109bd729e9ee91d620d1b883d87df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la habitación a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos que se pueden actualizar de la habitación",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "number": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "nightPrice": {
                                        "type": "integer"
                                    },
                                    "hotelId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Habitación actualizada"
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Eliminar una habitación",
                "description": "Elimina una habitación por su ID.",
                "operationId": "ba010fe020114b40d872de5e7913a8d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Habitación eliminada"
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    }
                }
            }
        },
...
    },
...
}